PPather: WmoFile: splitted into small classes!
